# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TiOS API",
        "description": "Welcolme to TiOS Platfor API documentation.",
        "termsOfService": "",
        "contact": {
            "email": "luis@tiosplatform.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000/api",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/datapoints": {
            "get": {
                "tags": [
                    "Datapoints"
                ],
                "summary": "Get list of datapoints",
                "description": "Returns list of datapoints",
                "operationId": "getDatapoints",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/datapoint/{id}": {
            "get": {
                "tags": [
                    "Datapoints"
                ],
                "summary": "Get datapoint information",
                "description": "Returns datapoint data",
                "operationId": "getDatapoint",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "datapoint id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/datapoints/types": {
            "get": {
                "tags": [
                    "Datapoints"
                ],
                "summary": "Get list of datapoints types",
                "description": "Returns list of datapoint types",
                "operationId": "getDatapointTypes",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/datapoint/datasource/{datasource_id}": {
            "get": {
                "tags": [
                    "Datapoints"
                ],
                "summary": "Get datapoints information by Datasource",
                "description": "Returns datapoints information",
                "operationId": "getDataPointByDataSource",
                "parameters": [
                    {
                        "name": "datasource_id",
                        "in": "path",
                        "description": "Datasource id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    },
...
}